api/routes.py
from flask import Blueprint, request, jsonify
import base64
import io
from PIL import Image

api_routes = Blueprint('api', __name__)

@api_routes.route('/upload', methods=['POST'])
def upload_files():
    files = request.files
    
    text_data = request.form.get('text', '')
    image_file = files.get('image')
    audio_file = files.get('audio')
    video_file = files.get('video')
    
    results = {
        "text_received": bool(text_data),
        "image_received": bool(image_file),
        "audio_received": bool(audio_file),
        "video_received": bool(video_file),
        "status": "files_uploaded"
    }
    
    return jsonify(results)

@api_routes.route('/batch_process', methods=['POST'])
def batch_process():
    data = request.json
    items = data.get('items', [])
    
    results = []
    for item in items:
        results.append({
            "id": item.get('id'),
            "status": "processed",
            "result": "sample_reasoning_output"
        })
    
    return jsonify({"results": results})