requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
opencv-python>=4.7.0
librosa>=0.10.0
numpy>=1.24.0
flask>=2.3.0
flask-cors>=4.0.0
scipy>=1.10.0
soundfile>=0.12.0